Separate archive processing and system profile parsing, reformat according to flake8
