+++ pyproject.toml
@@ -25,7 +25,7 @@ freezegun = "^0.3.15"
 
 [tool.poetry.scripts]
 puptoo = 'src.puptoo.app:main'
-puptoo-run = 'src.puptoo.process:run_profile'
+puptoo-run = 'src.puptoo.process.profile:run_profile'
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
